requirements.txt
asgiref==3.3.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
coloredlogs==15.0
Django==3.2
django-adminlte3==0.1.6
django-smart-selects==1.5.9
django-utils-six==2.0
djangorestframework==3.12.4
evernote==1.19
fastcore==1.3.20
humanfriendly==9.1
idna==3.3
img2pdf==0.4.0
lxml==4.6.3
packaging==20.9
pdf2image==1.14.0
pdfminer.six==20200517
pdfplumber==0.5.27
pikepdf==2.11.1
Pillow==8.1.2
pluggy==0.13.1
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyreadline==2.1
pytesseract==0.3.7
python-decouple==3.4
python-magic==0.4.13
python-magic-bin==0.4.14
pytz==2021.1
pywin32==300
reportlab==3.5.66
schedule==1.1.0
shortuuid==1.0.1
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
sweetify==2.0.0
tesseract==0.1.3
tqdm==4.60.0
urllib3==1.26.9
Wand==0.6.6
watchdog==2.0.2
wfastcgi==3.0.0
wincertstore==0.2
